school_management/requirements.txt
Django==4.2.23
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.6.0
django-redis==5.4.0
python-decouple==3.8
psycopg2-binary==2.9.10
Pillow==11.3.0
pandas==2.2.2
openpyxl==3.1.5
